8051-Sim900D-number-init-msg-sending
Upgraded code, now a user can store a mobile number to which message is to be sent before initializing the communication. 
Using # before a mobile number will tell the system to store it as a recipient number.
Using * before a message indicates that it is a message to be sent.